+++ store/db/mysql/memo_organizer.go
@@ -3,48 +3,31 @@ package mysql
 import (
 	"context"
 	"database/sql"
-	"fmt"
 	"strings"
 
 	"github.com/usememos/memos/store"
 )
 
 func (d *DB) UpsertMemoOrganizer(ctx context.Context, upsert *store.MemoOrganizer) (*store.MemoOrganizer, error) {
-	stmt := `
-		INSERT INTO memo_organizer (
-			memo_id,
-			user_id,
-			pinned
-		)
-		VALUES (?, ?, ?)
-		ON DUPLICATE KEY UPDATE pinned = ?
-	`
+	stmt := "INSERT INTO `memo_organizer` (`memo_id`, `user_id`, `pinned`) VALUES (?, ?, ?) ON DUPLICATE KEY UPDATE `pinned` = ?"
 	if _, err := d.db.ExecContext(ctx, stmt, upsert.MemoID, upsert.UserID, upsert.Pinned, upsert.Pinned); err != nil {
 		return nil, err
 	}
-
 	return upsert, nil
 }
 
 func (d *DB) GetMemoOrganizer(ctx context.Context, find *store.FindMemoOrganizer) (*store.MemoOrganizer, error) {
 	where, args := []string{}, []any{}
 	if find.MemoID != 0 {
-		where = append(where, "memo_id = ?")
+		where = append(where, "`memo_id` = ?")
 		args = append(args, find.MemoID)
 	}
 	if find.UserID != 0 {
-		where = append(where, "user_id = ?")
+		where = append(where, "`user_id` = ?")
 		args = append(args, find.UserID)
 	}
 
-	query := fmt.Sprintf(`
-		SELECT
-			memo_id,
-			user_id,
-			pinned
-		FROM memo_organizer
-		WHERE %s
-	`, strings.Join(where, " AND "))
+	query := "SELECT `memo_id`, `user_id`, `pinned` FROM `memo_organizer` WHERE " + strings.Join(where, " AND ")
 	row := d.db.QueryRowContext(ctx, query, args...)
 	if err := row.Err(); err != nil {
 		return nil, err
@@ -68,12 +51,12 @@ func (d *DB) GetMemoOrganizer(ctx context.Context, find *store.FindMemoOrganizer
 func (d *DB) DeleteMemoOrganizer(ctx context.Context, delete *store.DeleteMemoOrganizer) error {
 	where, args := []string{}, []any{}
 	if v := delete.MemoID; v != nil {
-		where, args = append(where, "memo_id = ?"), append(args, *v)
+		where, args = append(where, "`memo_id` = ?"), append(args, *v)
 	}
 	if v := delete.UserID; v != nil {
-		where, args = append(where, "user_id = ?"), append(args, *v)
+		where, args = append(where, "`user_id` = ?"), append(args, *v)
 	}
-	stmt := `DELETE FROM memo_organizer WHERE ` + strings.Join(where, " AND ")
+	stmt := "DELETE FROM `memo_organizer` WHERE " + strings.Join(where, " AND ")
 	if _, err := d.db.ExecContext(ctx, stmt, args...); err != nil {
 		return err
 	}
@@ -81,26 +64,10 @@ func (d *DB) DeleteMemoOrganizer(ctx context.Context, delete *store.DeleteMemoOr
 }
 
 func vacuumMemoOrganizer(ctx context.Context, tx *sql.Tx) error {
-	stmt := `
-	DELETE FROM
-		memo_organizer
-	WHERE
-		memo_id NOT IN (
-			SELECT
-				id
-			FROM
-				memo
-		)
-		OR user_id NOT IN (
-			SELECT
-				id
-			FROM
-				user
-		)`
+	stmt := "DELETE FROM `memo_organizer` WHERE `memo_id` NOT IN (SELECT `id` FROM `memo`) OR `user_id` NOT IN (SELECT `id` FROM `user`)"
 	_, err := tx.ExecContext(ctx, stmt)
 	if err != nil {
 		return err
 	}
-
 	return nil
 }
